- course management related comp have been moved to the designated folder - course-management comp works and implemented in the admin-dash
