+++ src/app/dashboard/admin-dash/course-management/student-data/student-data.component.ts
@@ -1,18 +1,18 @@
 import { Component, Input } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { combineLatest, map, Observable } from 'rxjs';
-import { Course } from '../../../core/user.model';
-import { SpinnerService } from '../../../core/services/spinner.service';
-import * as CourseSelectors from '../../../state/courses/course.selector';
-import * as UserSelectors from '../../../state/users/user.selector';
+import { Course } from '../../../../core/user.model';
+import { SpinnerService } from '../../../../core/services/spinner.service';
+import * as CourseSelectors from '../../../../state/courses/course.selector';
+import * as UserSelectors from '../../../../state/users/user.selector';
 import { AsyncPipe } from '@angular/common';
-import { SpinnerComponent } from '../../../core/spinner/spinner.component';
+import { SpinnerComponent } from '../../../../core/spinner/spinner.component';
 
 @Component({
   selector: 'app-student-data',
   imports: [AsyncPipe, SpinnerComponent],
   templateUrl: './student-data.component.html',
-  styleUrl: './student-data.component.css'
+  styleUrl: './student-data.component.css',
 })
 export class StudentDataComponent {
   @Input() courseId!: string;
@@ -21,31 +21,31 @@ export class StudentDataComponent {
   course$: Observable<Course | undefined>;
   students$: Observable<any[]>;
 
-  constructor(
-    private store: Store,
-    private spinner: SpinnerService
-  ) {
+  constructor(private store: Store, private spinner: SpinnerService) {
     this.course$ = this.store.select(CourseSelectors.selectCourseById(''));
     this.students$ = combineLatest([
       this.store.select(CourseSelectors.selectCourseById('')),
-      this.store.select(UserSelectors.selectAllUsers)
+      this.store.select(UserSelectors.selectAllUsers),
     ]).pipe(
       map(([course, users]) => {
         if (!course || !course.enrolledStudents) return [];
 
-        return course.enrolledStudents.map(studentId => {
-          const student = users.find(u => u.id === studentId);
+        return course.enrolledStudents.map((studentId) => {
+          const student = users.find((u) => u.id === studentId);
 
-          const grades = course.studentGrades && course.studentGrades[studentId]
-            ? [...course.studentGrades[studentId]]
-            : [];
+          const grades =
+            course.studentGrades && course.studentGrades[studentId]
+              ? [...course.studentGrades[studentId]]
+              : [];
 
           const attendanceMap = course.studentAttendance?.[studentId] || {};
-          const attendanceCount = Object.values(attendanceMap).filter(Boolean).length;
+          const attendanceCount =
+            Object.values(attendanceMap).filter(Boolean).length;
           const sessionsCount = course.sessions?.length || 0;
-          const attendanceRate = sessionsCount > 0
-            ? Math.round((attendanceCount / sessionsCount) * 100)
-            : 0;
+          const attendanceRate =
+            sessionsCount > 0
+              ? Math.round((attendanceCount / sessionsCount) * 100)
+              : 0;
 
           return {
             id: studentId,
@@ -55,10 +55,11 @@ export class StudentDataComponent {
             attendanceCount,
             sessionsCount,
             attendanceRate,
-            attendanceDetails: course.sessions?.map(session => ({
-              session,
-              present: attendanceMap[session.id] || false
-            })) || []
+            attendanceDetails:
+              course.sessions?.map((session) => ({
+                session,
+                present: attendanceMap[session.id] || false,
+              })) || [],
           };
         });
       })
@@ -67,27 +68,32 @@ export class StudentDataComponent {
 
   ngOnInit(): void {
     this.spinner.show();
-    this.course$ = this.store.select(CourseSelectors.selectCourseById(this.courseId));
+    this.course$ = this.store.select(
+      CourseSelectors.selectCourseById(this.courseId)
+    );
     this.students$ = combineLatest([
       this.store.select(CourseSelectors.selectCourseById(this.courseId)),
-      this.store.select(UserSelectors.selectAllUsers)
+      this.store.select(UserSelectors.selectAllUsers),
     ]).pipe(
       map(([course, users]) => {
         if (!course || !course.enrolledStudents) return [];
 
-        return course.enrolledStudents.map(studentId => {
-          const student = users.find(u => u.id === studentId);
+        return course.enrolledStudents.map((studentId) => {
+          const student = users.find((u) => u.id === studentId);
 
-          const grades = course.studentGrades && course.studentGrades[studentId]
-            ? [...course.studentGrades[studentId]]
-            : [];
+          const grades =
+            course.studentGrades && course.studentGrades[studentId]
+              ? [...course.studentGrades[studentId]]
+              : [];
 
           const attendanceMap = course.studentAttendance?.[studentId] || {};
-          const attendanceCount = Object.values(attendanceMap).filter(Boolean).length;
+          const attendanceCount =
+            Object.values(attendanceMap).filter(Boolean).length;
           const sessionsCount = course.sessions?.length || 0;
-          const attendanceRate = sessionsCount > 0
-            ? Math.round((attendanceCount / sessionsCount) * 100)
-            : 0;
+          const attendanceRate =
+            sessionsCount > 0
+              ? Math.round((attendanceCount / sessionsCount) * 100)
+              : 0;
 
           return {
             id: studentId,
@@ -97,10 +103,11 @@ export class StudentDataComponent {
             attendanceCount,
             sessionsCount,
             attendanceRate,
-            attendanceDetails: course.sessions?.map(session => ({
-              session,
-              present: attendanceMap[session.id] || false
-            })) || []
+            attendanceDetails:
+              course.sessions?.map((session) => ({
+                session,
+                present: attendanceMap[session.id] || false,
+              })) || [],
           };
         });
       })
@@ -122,7 +129,7 @@ export class StudentDataComponent {
     return new Date(date).toLocaleDateString('en-US', {
       month: 'short',
       day: 'numeric',
-      year: 'numeric'
+      year: 'numeric',
     });
   }
 }
